Add Percona Live Europ 2018 presentations
A smaller in scope, but updated to MySQL 8.0 and MariaDB 10.3
slide deck was created for the Percona Live Europe 2018 in Frankfurt.

Content for sql_modes, and other features was removed, but others
were expanded, as well as the theme was updated to Frankfurt and
de German Wiktionary. Topics like invisible indexes and DESC
indexes was included, but not newer features such as Common Table
Expressions and Window Functions.